Add script to fix ordering of scrambled PDBs and trajectories.
Such things can be generated by Desmond etc, which does not care
whether the atoms comprising a residue are next to each other in a
flat text file. There is nothing morally wrong with this but it makes
things inconvenient for the academic world, which still uses horrible
flat text FORTRAN friendly formats.